requirements.txt
# SatyaCheck AI - Production Requirements
# Optimized for Google Cloud Functions Python 3.11

# Core framework
functions-framework==3.5.0
flask>=2.3.3,<3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Google Cloud services (using latest compatible versions)
google-cloud-secret-manager==2.16.4
google-cloud-firestore==2.13.1
google-cloud-vision==3.4.5
google-cloud-translate==3.12.1
google-cloud-texttospeech==2.16.3
google-cloud-speech==2.21.0
google-cloud-pubsub==2.18.4
google-auth==2.23.4

# Google Gemini AI
google-generativeai==0.3.2

# Caching and performance
redis==5.0.1
hiredis==2.2.3

# Google API Client
google-api-python-client==2.108.0

# Web scraping and parsing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Image processing (compatible version for Python 3.11)
opencv-python-headless==4.8.1.78
numpy==1.25.2

# Enhanced HTTP handling
urllib3==2.0.7

# JSON processing and validation
jsonschema==4.19.2

# Security and validation (compatible versions)
cryptography>=40.0.0,<44.0.0
validators==0.22.0

# Enhanced text processing
regex==2023.10.3

# Enhanced datetime handling
python-dateutil==2.8.2

# Retry logic for robust API calls
tenacity==8.2.3

# Unicode handling for Indian languages
unidecode==1.3.7

# Audio/Video Processing for Deepfake Detection
opencv-python-headless==4.8.1.78
librosa==0.10.1
soundfile==0.12.1
ffmpeg-python==0.2.0

# Image processing for Gemini analysis
Pillow==10.1.0

# Twilio for WhatsApp & IVR integration
twilio==8.10.0